Add HomeFragment Top30 FragmentVeiwPager
