+++ app/src/main/java/com/song2/wave/UI/Main/Home/Top10/Idx3Fragment.kt
@@ -1,0 +1,18 @@
+package com.song2.wave.UI.Main.Home.Top10
+
+import android.os.Bundle
+import android.support.v4.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+
+class Idx3Fragment : Fragment(){
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
+        var v: View = inflater.inflate(com.song2.wave.R.layout.fragment_top10_idx3, container, false)
+
+        return v
+    }
+    override fun onActivityCreated(savedInstanceState: Bundle?) {
+        super.onActivityCreated(savedInstanceState)
+    }
+}
