+++ app/src/main/res/layout/fragment_top10.xml
@@ -1,0 +1,35 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+              android:orientation="vertical"
+              android:layout_width="match_parent"
+              android:layout_height="match_parent"
+              android:background="#000000"
+>
+
+    <RelativeLayout
+            android:layout_width="match_parent" android:layout_height="56dp">
+        <TextView
+                android:layout_centerInParent="true"
+                android:fontFamily="@font/noto_sans_kr_bold_hestia"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:textSize="18sp"
+                android:textColor="#ffffff"
+                android:lineSpacingExtra="6sp"
+                android:text="TOP 10"
+        />
+    </RelativeLayout>
+
+    <android.support.design.widget.TabLayout
+            android:id="@+id/tl_top10_frag_tabbar"
+            android:layout_width="match_parent"
+            android:layout_height="52dp"/>
+
+    <android.support.v4.view.ViewPager
+            android:id="@+id/vp_top10_frag_content"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"/>
+
+
+</LinearLayout>
