+++ app/src/main/res/layout/fragment_top10_tabbar.xml
@@ -1,0 +1,125 @@
+<?xml version="1.0" encoding="utf-8"?>
+<HorizontalScrollView
+        android:layout_width="match_parent"
+        android:layout_height="50dp"
+        xmlns:android="http://schemas.android.com/apk/res/android">
+    <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+                  android:layout_width="match_parent"
+                  android:layout_height="match_parent"
+                  android:background="#000000"
+    >
+        <RelativeLayout
+                android:id="@+id/rl_top10_frag_idx0"
+                android:layout_width="65dp"
+                android:layout_height="50dp"
+                android:gravity="center">
+            <TextView
+                    android:text="idx0"
+                    android:gravity="center"
+                    android:textSize="16sp"
+                    android:layout_width="wrap_content"
+                    android:layout_height="match_parent"
+                    android:fontFamily="@font/noto_sans_kr_bold_hestia"
+                    android:includeFontPadding="false"/>
+
+        </RelativeLayout>
+        <RelativeLayout
+                android:id="@+id/rl_top10_frag_idx1"
+                android:layout_width="65dp"
+                android:layout_height="match_parent"
+                android:gravity="center">
+            <TextView
+                    android:text="idx1"
+                    android:gravity="center"
+                    android:textSize="16sp"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:fontFamily="@font/noto_sans_kr_bold_hestia"
+                    android:includeFontPadding="false"/>
+        </RelativeLayout>
+        <RelativeLayout
+                android:id="@+id/rl_top10_frag_idx2"
+                android:layout_width="65dp"
+                android:layout_height="match_parent"
+                android:gravity="center">
+            <TextView
+                    android:text="idx2"
+                    android:textSize="16sp"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:fontFamily="@font/noto_sans_kr_bold_hestia"
+                    android:includeFontPadding="false"
+            />
+        </RelativeLayout>
+        <RelativeLayout
+                android:id="@+id/rl_top10_frag_idx3"
+                android:layout_width="65dp"
+                android:layout_height="match_parent"
+                android:gravity="center">
+            <TextView
+                    android:text="idx3"
+                    android:textSize="16sp"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:fontFamily="@font/noto_sans_kr_bold_hestia"
+                    android:includeFontPadding="false"
+            />
+
+        </RelativeLayout>
+        <RelativeLayout
+                android:id="@+id/rl_top10_frag_idx4"
+                android:layout_width="65dp"
+                android:layout_height="match_parent"
+                android:gravity="center">
+            <TextView
+                    android:text="idx4"
+                    android:textSize="16sp"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:fontFamily="@font/noto_sans_kr_bold_hestia"
+                    android:includeFontPadding="false"/>
+        </RelativeLayout>
+        <RelativeLayout
+                android:id="@+id/rl_top10_frag_idx5"
+                android:layout_width="65dp"
+                android:layout_height="match_parent"
+                android:gravity="center">
+            <TextView
+                    android:text="idx5"
+                    android:textSize="16sp"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:fontFamily="@font/noto_sans_kr_bold_hestia"
+                    android:includeFontPadding="false"
+            />
+        </RelativeLayout>
+        <RelativeLayout
+                android:id="@+id/rl_top10_frag_idx6"
+                android:layout_width="65dp"
+                android:layout_height="match_parent"
+                android:gravity="center">
+            <TextView
+                    android:text="idx6"
+                    android:textSize="16sp"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:fontFamily="@font/noto_sans_kr_bold_hestia"
+                    android:includeFontPadding="false"
+            />
+
+        </RelativeLayout>
+        <RelativeLayout
+                android:id="@+id/rl_top10_frag_idx7"
+                android:layout_width="65dp"
+                android:layout_height="match_parent"
+                android:gravity="center">
+            <TextView
+                    android:text="idx7"
+                    android:textSize="16sp"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:fontFamily="@font/noto_sans_kr_bold_hestia"
+                    android:includeFontPadding="false"/>
+        </RelativeLayout>
+    </LinearLayout>
+</HorizontalScrollView>
